+++ src/RcppFunctions.cpp
@@ -412,3 +412,5 @@ bool set_proj_search_paths(std::vector<std::string> paths) {
 }
 
 
+
+