+++ src/raster_methods.cpp
@@ -1496,7 +1496,7 @@ SpatRaster SpatRaster::init(std::string value, bool plusone, SpatOptions &opt) {
 	size_t nc = ncol();
 	std::vector<double> v;
 	if (value == "row") {
-		for (size_t i = 0; i < out.bs.n; i++) {
+		for (size_t i = 0; i < out.bs.n; i++) {	
 			v.resize(nc * out.bs.nrows[i]);
 			for (size_t j = 0; j < out.bs.nrows[i]; j++) {
 				size_t r = out.bs.row[i] + j + plusone;
@@ -1506,6 +1506,9 @@ SpatRaster SpatRaster::init(std::string value, bool plusone, SpatOptions &opt) {
 			}
 			if (!out.writeValues(v, out.bs.row[i], out.bs.nrows[i], 0, nc)) return out;
 		}
+		source[0].range_min.resize(1, 0 + plusone);
+		source[0].range_max.resize(1, nrow() - 1 + plusone);
+		source[0].hasRange.resize(1, true);
 	} else if (value == "col") {
 		std::vector<double> cnn(nc);
 		double start = plusone ? 1 : 0;
@@ -1518,6 +1521,9 @@ SpatRaster SpatRaster::init(std::string value, bool plusone, SpatOptions &opt) {
 			}
 			if (!out.writeValues(v, out.bs.row[i], out.bs.nrows[i], 0, nc)) return out;
 		}
+		source[0].range_min.resize(1, 0 + plusone);
+		source[0].range_max.resize(1, nc - 1 + plusone);
+		source[0].hasRange.resize(1, true);
 	} else if (value == "cell") {
 		for (size_t i = 0; i < out.bs.n; i++) {
 			v.resize(nc * out.bs.nrows[i]);
@@ -1526,6 +1532,10 @@ SpatRaster SpatRaster::init(std::string value, bool plusone, SpatOptions &opt) {
 			std::iota(v.begin(), v.end(), firstcell);
 			if (!out.writeValues(v, out.bs.row[i], out.bs.nrows[i], 0, nc)) return out;
 		}
+		source[0].range_min.resize(1, 0 + plusone);
+		source[0].range_max.resize(1, ncell() - 1 + plusone);
+		source[0].hasRange.resize(1, true);
+		
 	} else if (value == "x") {
 		std::vector<int_64> col(nc);
 		std::iota(col.begin(), col.end(), 0);
@@ -1538,6 +1548,10 @@ SpatRaster SpatRaster::init(std::string value, bool plusone, SpatOptions &opt) {
 			}
 			if (!out.writeValues(v, out.bs.row[i], out.bs.nrows[i], 0, nc)) return out;
 		}
+		source[0].range_min.resize(1, xcoords[0]);
+		source[0].range_max.resize(1, xcoords[nc-1]);
+		source[0].hasRange.resize(1, true);
+
 	} else if (value == "y") {
 	
 		for (size_t i = 0; i < out.bs.n; i++) {
@@ -1550,6 +1564,9 @@ SpatRaster SpatRaster::init(std::string value, bool plusone, SpatOptions &opt) {
 			}
 			if (!out.writeValues(v, out.bs.row[i], out.bs.nrows[i], 0, nc)) return out;
 		}
+		source[0].range_min.resize(1, yFromRow(0));
+		source[0].range_max.resize(1, yFromRow(nrow()-1));
+		source[0].hasRange.resize(1, true);
 
 	} else if (value == "chess") {
 		std::vector<double> a(nc);
@@ -1571,6 +1588,9 @@ SpatRaster SpatRaster::init(std::string value, bool plusone, SpatOptions &opt) {
 			recycle(v, out.bs.nrows[i] * nc);
 			if (!out.writeValues(v, out.bs.row[i], out.bs.nrows[i], 0, nc)) return out;
 		}
+		source[0].range_min.resize(1, 0);
+		source[0].range_max.resize(1, 1);
+		source[0].hasRange.resize(1, true);
 	}
 
 	out.writeStop();
@@ -1590,6 +1610,10 @@ SpatRaster SpatRaster::init(std::vector<double> values, SpatOptions &opt) {
 			v.resize(out.bs.nrows[i]*nc*nl, val);				
 			if (!out.writeValues(v, out.bs.row[i], out.bs.nrows[i], 0, nc)) return out;
 		}
+		source[0].range_min.resize(1, val);
+		source[0].range_max.resize(1, val);
+		source[0].hasRange.resize(1, true);
+		
 	} else {
 		int over = 0;
 		for (size_t i = 0; i < out.bs.n; i++) {